+++ src/tm_devices/drivers/device.py
@@ -73,7 +73,9 @@ class Device(_AbstractDeviceControl, _ExtendableMixin, ABC):
 
         Args:
             config_entry: A config entry object parsed by the DMConfigParser.
-            verbose: A boolean indicating if verbose output should be printed.
+            verbose: A boolean indicating if verbose output should be printed. If True,
+                communication printouts will be logged with a level of INFO. If False,
+                communication printouts will be logged with a level of DEBUG.
         """
         super().__init__(config_entry, verbose)
         self._is_open = True